requirements.txt
# Core Framework
Flask==2.3.3
Werkzeug==2.3.7

# Machine Learning & AI
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3

# Image Processing
Pillow==10.0.1
opencv-python-headless==4.8.1.78

# Data Science & Visualization
matplotlib==3.7.2
pandas==2.0.3

# Web Scraping & Research
requests==2.31.0
beautifulsoup4==4.12.2

# Additional Utilities
tqdm==4.66.1
gunicorn==21.2.0